[WaveformEditor] Logarithmic Zoom #trivial #rb none #preflight 6319c751980179553ec0bba3
[CL 21891959 by igor dallavanzi in ue5-main branch]

+++ Engine/Plugins/Editor/WaveformEditor/Source/WaveformEditorWidgets/Public/WaveformEditorTransportCoordinator.h
@@ -40,7 +40,7 @@ public:
 
 	const float GetPlayheadPosition() const;
 	void ReceivePlayBackRatio(const float NewRatio);
-	void OnZoomLevelChanged(const uint8 NewLevel);
+	void OnZoomLevelChanged(const float NewLevel);
 
 	float ConvertAbsoluteRatioToZoomed(const float InAbsoluteRatio) const;
 	float ConvertZoomedRatioToAbsolute(const float InZoomedRatio) const;
